+++ backends/bevy_picking_rapier/src/lib.rs
@@ -93,8 +93,7 @@ pub fn update_hits(
             continue;
         }
 
-        // Cameras missing render layers intersect all layers
-        let cam_layers = cam_layers.copied().unwrap_or_else(RenderLayers::all);
+        let cam_layers = cam_layers.copied().unwrap_or_default();
 
         let predicate = |entity| {
             let marker_requirement =